Recast the JWE key-recovery path as key-owning decryptor composition
The external-key encryption path was an optional IExternalKeyEncryptor port
threaded through JsonWebTokenEncryptor with the local-vs-custodian routing inlined.
Recovering the CEK is the only private operation, so it moves behind a new
IDataDecryptor seam that mirrors the signing seam: LocalKeyDecryptor and
ExternalKeyDecryptor are peer backends owning private-bearing and public-only keys,
and CompositeDataDecryptor routes each recovery to the owner, yielding a uniform
null (RFC 7516 section 11.5) when none owns the key. Producing a JWE wraps the CEK
with the recipient's public half or a local secret, so it never needs a custodian
and stays in JsonWebTokenEncryptor, exactly as signature verification stays out of
the signing seam.

The custodian contract slims to IKeyCustodian (UnwrapKeyAsync, AgreeKeyAsync); the
dead Seal/Open methods and the symmetric wrap operation, which no flow reaches (the
server encrypts outbound with recipient keys, never its own), are dropped. The
startup ExternalKeyWiringValidator is removed, its checks now covered by the runtime
fail-closed path, mirroring the signing seam's earlier removal.
